Added intellisense to rich textbox
The story board text box now has intellisense from Rabbit.Controls
thanks to
(http://www.codeproject.com/Articles/180311/Rich-Text-Box-With-Intellisense-Ability).
It was really helpful understanding more of the WPF controls and how to
bind them to code from the view model. A concept that is still hard to
grasp in WPF.